Move components one level up in `src/sidebar-common/components`.
