+++ src/sidebar-common/components/annotation-highlight.css
