+++ src/sidebar-common/components/annotation-highlight.tsx
