--- src/sidebar-common/components/annotation-highlight/index.ts
@@ -1,3 +1,0 @@
-import AnnotationHighlight from './annotation-highlight'
-
-export default AnnotationHighlight
