+++ src/sidebar-common/components/close-button.css
