--- src/sidebar-common/components/close-button/index.tsx
@@ -1,3 +1,0 @@
-import CloseButton from './close-button'
-
-export default CloseButton
