+++ src/sidebar-common/components/congrats-message.css
