+++ src/sidebar-common/components/congrats-message.tsx
