--- src/sidebar-common/components/congrats-message/index.tsx
@@ -1,3 +1,0 @@
-import CongratsMessage from './congrats-message'
-
-export default CongratsMessage
