+++ src/sidebar-common/components/empty-message.css
