+++ src/sidebar-common/components/empty-message.tsx
