--- src/sidebar-common/components/empty-message/index.tsx
@@ -1,3 +1,0 @@
-import EmptyMessage from './empty-message'
-
-export default EmptyMessage
