+++ src/sidebar-common/components/footer.css
