+++ src/sidebar-common/components/footer.tsx
