--- src/sidebar-common/components/footer/index.tsx
@@ -1,3 +1,0 @@
-import Footer from './footer'
-
-export default Footer
