+++ src/sidebar-common/components/loader.css
