+++ src/sidebar-common/components/loader.tsx
