--- src/sidebar-common/components/loader/index.tsx
@@ -1,3 +1,0 @@
-import Loader from './loader'
-
-export default Loader
