+++ src/sidebar-common/components/tag-holder.css
