--- src/sidebar-common/components/tag-holder/index.tsx
@@ -1,3 +1,0 @@
-import TagHolder from './tag-holder'
-
-export default TagHolder
