+++ src/sidebar-common/components/topbar.css
