--- src/sidebar-common/components/topbar/index.tsx
@@ -1,3 +1,0 @@
-import Topbar from './topbar'
-
-export default Topbar
