+++ src/sidebar-common/components/truncated-text-renderer.css
