+++ src/sidebar-common/components/truncated-text-renderer.tsx
