--- src/sidebar-common/components/truncated-text-renderer/index.ts
@@ -1,3 +1,0 @@
-import TruncatedTextRenderer from './truncated-text-renderer'
-
-export default TruncatedTextRenderer
